backend/requirements.txt
fastapi
uvicorn
websockets
google-genai
pydantic
python-dotenv
google-adk
